+++ src/lib/libwasm_worker.js
@@ -39,6 +39,9 @@
 #endif
 `;
   const wasmWorkerOptions = `{
+#if EXPORT_ES6
+  'type': 'module',
+#endif
 #if ENVIRONMENT_MAY_BE_NODE
   // This is the way that we signal to the node worker that it is hosting
   // a wasm worker.